Исправил корректность выбора оферов в форме
