+++ src/utils-constant/utils.js
@@ -22,23 +22,6 @@ function capitalizeFirstLetter(text){
   return text.charAt(0).toUpperCase() + text.slice(1);
 }
 
-const createOfferItemTemplate = (type, title, price, className) => `
-  <div class="event__offer-selector">
-    <input class="event__offer-checkbox  visually-hidden" id="event-offer-${type}-1" type="checkbox" name="event-offer-${type}" ${className}>
-    <label class="event__offer-label" for="event-offer-${type}-1">
-      <span class="event__offer-title">${title}</span>
-      &plus;&euro;&nbsp;
-      <span class="event__offer-price">${price}</span>
-    </label>
-  </div>
-`;
-
-const createTypeGroupTemplate = (group, className) => `
-  <div class="event__type-item">
-    <input id="event-type-${group.toLowerCase()}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${group.toLowerCase()}" ${className}>
-    <label class="event__type-label  event__type-label--${group.toLowerCase()}" for="event-type-${group.toLowerCase()}-1">${group}</label>
-  </div>
-`;
 
 function updateItem(items, update) {
   return items.map((item) => item.id === update.id ? update : item);
@@ -53,6 +36,4 @@ export {
   getRandomArrayElement,
   getRandomInteger,
   capitalizeFirstLetter,
-  createOfferItemTemplate,
-  createTypeGroupTemplate
 };
