+++ src/view/editor-point-view.js
@@ -2,7 +2,7 @@ import AbstractStatefulView from '../framework/view/abstract-stateful-view.js';
 import flatpickr from 'flatpickr';
 import 'flatpickr/dist/flatpickr.min.css';
 import { TYPES } from '../utils-constant/constant.js';
-import { createOfferItemTemplate, createTypeGroupTemplate } from '../utils-constant/utils.js';
+//import { createTypeGroupTemplate } from '../utils-constant/utils.js';
 
 function createPicturesTemplate(photo) {
   return photo.reduce((acc, {src}) => {
@@ -13,6 +13,13 @@ function createPicturesTemplate(photo) {
   }, '');
 }
 
+const createTypeGroupTemplate = (group, className) => `
+  <div class="event__type-item">
+    <input id="event-type-${group.toLowerCase()}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${group.toLowerCase()}" ${className}>
+    <label class="event__type-label  event__type-label--${group.toLowerCase()}" for="event-type-${group.toLowerCase()}-1">${group}</label>
+  </div>
+`;
+
 function createAddPointTemplate(state, allDestination) {
   const { basePrice, type, offers, typeOffers, destination } = state;
   // { basePrice, type } = point;
@@ -20,11 +27,22 @@ function createAddPointTemplate(state, allDestination) {
   const pointDestination = allDestination.find((item) => item.id === destination);
   const { name, description, pictures } = pointDestination;
 
-  const createAllOffers = typeOffers.offers
-    .map((offer) => {
-      const checkedClassName = offers.includes(offer.id) ? 'checked' : '';
-      return createOfferItemTemplate(type, offer.title, offer.price, offer.id, checkedClassName);
-    }).join('');
+  const createOfferItemTemplate = (offerItem, isCheckedOfferItem) =>
+    `
+      <div class="event__offer-selector">
+        <input class="event__offer-checkbox visually-hidden" id="event-offer-${offerItem.id}-1" type="checkbox" name="event-offer-${type.toLowerCase()}" value="${offerItem.id}" ${isCheckedOfferItem ? 'checked' : ''}>
+        <label class="event__offer-label" for="event-offer-${offerItem.id}-1">
+          <span class="event__offer-title">${offerItem.title}</span>
+          &plus;&euro;&nbsp;
+          <span class="event__offer-price">${offerItem.price}</span>
+        </label>
+      </div>
+    `;
+
+  const createAllOffersTemplate = typeOffers.offers.map((offerItem) => {
+    const isCheckedOfferItem = offers.includes(offerItem.id);
+    return createOfferItemTemplate(offerItem, isCheckedOfferItem);
+  }).join('');
 
   const createDestinationTemplate = allDestination
     .map((item) => `<option value="${item.name}"></option>`).join('');
@@ -85,7 +103,7 @@ function createAddPointTemplate(state, allDestination) {
         <section class="event__section  event__section--offers">
           <h3 class="event__section-title  event__section-title--offers">Offers</h3>
           <div class="event__available-offers">
-            ${createAllOffers}
+            ${createAllOffersTemplate}
           </div>
         </section>
         <section class="event__section  event__section--destination">
@@ -113,14 +131,14 @@ export default class EditorPointView extends AbstractStatefulView{
   #handleEditRollUp = null;
   #datepickerStart = null;
   #datepickerEnd = null;
-  constructor({point, typeOffers, allOffers, pointDestination, allDestination, onFormSubmit, onEditRollup, onCloseEditButtonClick}) {
+  constructor({point, typeOffers, allOffers, pointDestination, allDestination, onFormSubmit, onCloseEditButtonClick}) {
     super();
     this.#point = point;
     this.#allOffers = allOffers;
     this.#allDestination = allDestination;
     this.#pointDestination = pointDestination;
     this.#handleFormSubmit = onFormSubmit;
-    this.#handleEditRollUp = onEditRollup;
+    //this.#handleEditRollUp = onEditRollup;
     this.#onCloseEditButtonClick = onCloseEditButtonClick;
     this._setState(EditorPointView.parsePointToState(point, pointDestination.id, typeOffers));
     this._restoreHandlers();
@@ -185,6 +203,16 @@ export default class EditorPointView extends AbstractStatefulView{
     this._setState({
       dateFrom: userDate,
     });
+  };
+
+  #closeEditButtonClickHandler = (evt) => {
+    evt.preventDefault();
+    this.#onCloseEditButtonClick();
+  };
+
+  #submitButtonClickHandler = (evt) => {
+    evt.preventDefault();
+    this.#onSubmitButtonClick(this.#point);
   };
 
   #dateToChangeHandler = ([userDate]) => {
